requirements.txt
flask
requests
pillow
python-magic
